# backend/internal/models/recommendation.go
package models

// Recommendation represents advice generated by the decision engine.
type Recommendation struct {
	ID        int    `json:"id"`
	FarmID    int    `json:"farm_id"`
	Message   string `json:"message"`
	Priority  string `json:"priority"`
	Reason    string `json:"reason"`
	CreatedAt string `json:"created_at"`
}